Add checking sample names in get_counts
get_counts now checks whether there are samples missing from the data that
are in the supplied sample object. This causes an error.
It also checks if there are samples in the data that aren't in the supplied
samples object. This generates a warning to allow for subsetting.
A test file has alos been created to test get_counts, get_gene_metadata and
subset_to_samples.